[REFACTOR] common aggregate 추가 및 리팩토링 + @SequenceGenerator 삭제
