+++ team/src/main/java/sansam/team/project/command/domain/aggregate/entity/Project.java
@@ -3,7 +3,7 @@ package sansam.team.project.command.domain.aggregate.entity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sansam.team.common.BaseTimeEntity;
+import sansam.team.common.aggregate.entity.BaseTimeEntity;
 import sansam.team.project.command.application.dto.project.ProjectUpdateDTO;
 import sansam.team.project.command.domain.aggregate.ProjectStatus;
