+++ team/src/main/java/sansam/team/team/command/domain/aggregate/entity/Team.java
@@ -2,22 +2,17 @@ package sansam.team.team.command.domain.aggregate.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sansam.team.common.BaseTimeEntity;
+import sansam.team.common.aggregate.entity.BaseTimeEntity;
 
 @Entity
 @Table(name = "tbl_team")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@SequenceGenerator(
-        name = "TEAM_SEQ_GENERATOR"
-        , sequenceName = "team_seq" //매핑할 데이터베이스 시퀀스 이름
-        , allocationSize = 1)
 public class Team extends BaseTimeEntity {
 
-    @Id @GeneratedValue(strategy = GenerationType.AUTO, generator = "TEAM_SEQ_GENERATOR")
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long teamSeq;   // 팀 번호
     private long projectSeq;    // 프로젝트 번호
     private long ruleSeq;   // 팀 빌딩 규칙 번호
