+++ team/src/main/java/sansam/team/team/command/domain/aggregate/entity/TeamMember.java
@@ -4,8 +4,8 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sansam.team.common.BaseTimeEntity;
-import sansam.team.common.YnType;
+import sansam.team.common.aggregate.entity.BaseTimeEntity;
+import sansam.team.common.aggregate.YnType;
 import sansam.team.team.command.domain.aggregate.InterestType;
 
 @Entity
@@ -18,7 +18,7 @@ import sansam.team.team.command.domain.aggregate.InterestType;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class TeamMember extends BaseTimeEntity {
 
-    @Id @GeneratedValue(strategy = GenerationType.AUTO, generator = "TEAM_MEMBER_SEQ_GENERATOR")
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long teamMemberSeq;
     private long teamSeq;
     private long userSeq;
