+++ team/src/main/java/sansam/team/team/command/domain/aggregate/entity/TeamMemberScheduleProgress.java
@@ -4,19 +4,15 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sansam.team.common.BaseTimeEntity;
+import sansam.team.common.aggregate.entity.BaseTimeEntity;
 
 @Entity
 @Table(name = "tbl_team_member_schedule_progress")
-@SequenceGenerator(
-        name = "TEAM_MEMBER_SCHEDULE_PROGRESS_SEQ_GENERATOR"
-        , sequenceName = "TEAM_MEMBER_SCHEDULE_PROGRESS_SEQ"
-        , allocationSize = 1)
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class TeamMemberScheduleProgress extends BaseTimeEntity {
 
-    @Id @GeneratedValue(strategy = GenerationType.AUTO, generator = "TEAM_MEMBER_SCHEDULE_PROGRESS_SEQ_GENERATOR")
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long teamMemberScheduleProgressSeq;
     private long teamScheduleSeq;
     private long teamMemberSeq;
