+++ team/src/main/java/sansam/team/user/query/dto/UserQueryDTO.java
@@ -1,9 +1,9 @@
 package sansam.team.user.query.dto;
 
 import lombok.*;
-import sansam.team.common.BaseTimeEntity;
+import sansam.team.common.aggregate.entity.BaseTimeEntity;
 import sansam.team.common.jwt.JwtToken;
-import sansam.team.user.command.domain.aggregate.RoleType;
+import sansam.team.common.aggregate.RoleType;
 import sansam.team.user.command.domain.aggregate.StatusType;
 
 import java.time.LocalDateTime;
